Fix issue with Futures created from Promises
This fix makes it so when the `fork`, `value`, `done`, or `promise`
functions throw an exception as a result of the Future entering a
crashed state, the exception is thrown in the next tick.

This means that if a Promise was used anywhere in the pipeline, we can
ensure that it does not get the opportunity to catch this exception.

The reason we want to avoid this, is because the Promise continuations
are already handled by Fluture, and if the Promise catches this
exception, it will start handling it in its own way, by emitting an
unhandledRejection event on some global event emitter.