Expand the section on type signatures in the README
The section was removed after switching to simple currying, as most of
it had become inaccurate. However, since we link to it with "How to read
the type signatures", we should at least have some content there for new
users. Simply having a short introduction to Hindley-Milner in general
should do it.