Add toJSON methods to Fluture's data structures
Since casting a Future to JSON doesn't really make sense, I've decided
to implement it as returning a kind of AST. This AST can be useful for
debugging. Since most of the scenarios where a Future is being turned
into JSON is during error handling, returning a value useful for
debugging seemed the most appropriate.