Rework tests that relied on async throw for assertions
After switching to Oletus in #368 (6868f56), tests that throw their
assertion error asynchronously no longer report their errors in a
readable way.

Mocha would catch these errors, and, as Mocha only ever runs a single
test at once, would know which test to fail. In the case of Oletus, the
whole test suite fails when a single test throws an async error, and as
such, the source of this error is lost.

It was very easy to move from throwing these error to just passing them
into the done() function, so that's what this commit does.