Reduce likelyhood of false positives in parallel tests
After switching to Oletus in #368 (6868f56), all tests now run in
parallel. This means that timing-sensitive tests have a greater chance
of producing a false positive due to reordering in the event loop.

We're reducing the risk of this happening by simply adding more time
between the events that must occur in a specific ordering.