Change all file extensions back to .js
The Node 12 experimental module system allows files to have regular .js
extensions, which are interpreted as modules as long as the "closest"
package.json contains '"type": "module"'.

The reason why I'm moving to this approach, over explicitly naming files
with the .mjs extension, is so that outdated tools such as esm, nyc,
ts-node, etc. that just-in-time compile the code to a CJS format and
attempt to require it, don't get Node 12 to throw an error because the
file extension doesn't match the module type.